requirements.txt
discord.py==1.2.4
python-dotenv==0.10.3
